+++ src/main/java/com/alkemy/ong/domain/categories/Category.java
@@ -1,4 +1,4 @@
-package com.alkemy.ong.domain.category;
+package com.alkemy.ong.domain.categories;
 
 import lombok.Builder;
 import lombok.Getter;
