+++ src/main/java/com/alkemy/ong/domain/categories/CategoryService.java
@@ -1,4 +1,4 @@
-package com.alkemy.ong.domain.category;
+package com.alkemy.ong.domain.categories;
 
 import org.springframework.stereotype.Service;
