+++ src/main/java/com/alkemy/ong/web/controllers/CategoryController.java
@@ -1,7 +1,7 @@
 package com.alkemy.ong.web.controllers;
 
-import com.alkemy.ong.domain.category.Category;
-import com.alkemy.ong.domain.category.CategoryService;
+import com.alkemy.ong.domain.categories.Category;
+import com.alkemy.ong.domain.categories.CategoryService;
 import lombok.Builder;
 import lombok.Data;
 import org.springframework.http.HttpStatus;
